// src/app/models/pnr/emd.model.ts
export class EmdModel {
    id: number;
    emdType: string;
    ticketNumber: string;
    baseAmount: string;
    currency: string;
    otherTaxes: string;
    taxes: string;
    totalCurrency: string;
    totalCharge: string;
    fop: string;
    invoiceNumber: string;
    carrierCode: string;
    passengerName: string;
}